# forgepay/api/openapi.yaml
openapi: "3.1.0"
info:
  title: ForgePay API
  version: "2026-04"
  description: |
    ForgePay unified payment API — accept card, stablecoin, and crypto payments;
    manage subscriptions; and query events. All amounts are integers in the
    smallest currency unit (cents for USD/EUR) unless noted.
  contact:
    name: ForgePay Developer Support
    url: https://myforgepay.com/docs
    email: api@myforgepay.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.myforgepay.com
    description: Production
  - url: https://api.sandbox.myforgepay.com
    description: Sandbox
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: "Use your ForgePay API key: Authorization: Bearer sk_live_..."
  schemas:
...
    Payment:
      type: object
      properties:
        id:             { type: string, example: "pay_01HXYZ..." }
        amount:         { type: integer, example: 4900 }
        currency:       { type: string, example: "USD" }
        status:         { type: string, enum: [created, processing, succeeded, failed, cancelled, refunded] }
        customer_id:    { type: string }
        client_secret:  { type: string }
        payment_method: { type: string, enum: [card, bank_transfer, wallet] }
        created_at:     { type: string, format: date-time }
        metadata:       { type: object }
      required: [id, amount, currency, status]
...
    Error:
      type: object
      properties:
        error:   { type: string }
        message: { type: string }
        code:    { type: string }
      required: [error, message]
    ListMeta:
      type: object
      properties:
        has_more:    { type: boolean }
        total_count: { type: integer }
        next_cursor: { type: string }
    Agent:
      type: object
      properties:
        id:               { type: string, example: "agt_01HXYZ..." }
        name:             { type: string, example: "logistics-bot-v1" }
        capabilities:     { type: array, items: { type: string } }
        reputation_score: { type: integer, minimum: 0, maximum: 100, example: 72 }
        did:              { type: string, description: "Decentralized Identifier" }
        status:           { type: string, enum: [active, suspended, deregistered] }
        created_at:       { type: string, format: date-time }
        metadata:         { type: object }
      required: [id, name, capabilities, reputation_score, status]
...
paths:
  /v1/payments:
    post:
      summary: Create a payment
      operationId: createPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:          { type: integer, example: 4900 }
                currency:        { type: string, example: "USD" }
                customer_id:     { type: string }
                payment_method_data: { type: object }
                idempotency_key: { type: string }
                metadata:        { type: object }
              required: [amount, currency]
      responses:
        "200": { description: Payment created, content: { application/json: { schema: { $ref: "#/components/schemas/Payment" } } } }
        "400": { description: Bad request, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: Rate limit exceeded }
    get:
      summary: List payments
      operationId: listPayments
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 20, maximum: 100 } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: customer_id, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string } }
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ListMeta"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Payment" }
...
  /v1/agents/{agentId}:
    get:
      summary: Get agent by ID
      operationId: getAgent
      parameters:
        - { name: agentId, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Agent, content: { application/json: { schema: { $ref: "#/components/schemas/Agent" } } } }
        "404": { description: Not found }
...